This code was downloaded from Apple, version 61. https://opensource.apple.com/tarballs/OpenAL/OpenAL-61.tar.gz
I already started hacking on it to get it compiling.
This also required the CoureAudioUtilityClasses which I embedded
directly in the project.
https://developer.apple.com/library/archive/samplecode/CoreAudioUtilityClasses/Introduction/Intro.html#//apple_ref/doc/uid/DTS40012328
https://developer.apple.com/library/archive/samplecode/CoreAudioUtilityClasses/CoreAudioUtilityClasses.zip

The Mac Xcode project is now fixed. The Deployment target is set to 10.9
because anything less needs to link in a legacy version of the standard
C++ library, and I didn't feel like mucking with the build flags or
testing across all the macOS versions to make sure it actually works. I
also changed the install paths and runtime paths so it can be bundled.

I started a new iOS project. It does not build yet. I started hacking in
commited first. Right now, the build fails in the Capture implementation
because it appears to use Carbon. I recall iOS originally did not
implement capture on iOS and for many years did not implement. I have
not checked since. But I suspect the easiest thing to do is stub out all
the capture stuff for iOS and hope the rest works.